Update the tests and test infrastructure in test/regression
Tests updated to work with x86 64 bits.
Infrastructure added: script "Runtest", with ability to have different reference outputs depending on platform or bit size.

+++ test/regression/sizeof2.c
@@ -3,7 +3,7 @@
 
 int main()
 {
-  printf("%d\n", sizeof("abcd"));
-  printf("%d\n", sizeof(L"abcd") / sizeof(wchar_t));
+  printf("%d\n", (int) sizeof("abcd"));
+  printf("%d\n", (int) (sizeof(L"abcd") / sizeof(wchar_t)));
   return 0;
 }
